refactored code so filtered friends is available at a common ancestor
