+++ src/components/friends/FriendsProvider.js
@@ -3,7 +3,7 @@
  *  PURPOSE: This module provides FRIENDS data to the app.
  **/
 
-import React, { useState, createContext } from "react"
+import React, { useState, createContext, useEffect } from "react"
 
 export const FriendsContext = createContext()
 
@@ -11,6 +11,8 @@ export const FriendsProvider = (props) => {
 
     // Manages the state of all Friends realtionships from the database
     const [friends, setFriends] = useState([])
+    // Manages state of friends of the current user
+    const [filteredFriends, setFilteredFriends] = useState([])
 
     const getFriends = () => {
         return fetch("http://localhost:8088/friends?_expand=user")
@@ -18,9 +20,17 @@ export const FriendsProvider = (props) => {
             .then(setFriends)
     }
 
+    const filterFriends = () => {
+        return friends.filter(friend => friend.currentUserId === parseInt(sessionStorage.nutshell_user))
+    }
+
+    useEffect(() => {
+        setFilteredFriends(filterFriends())
+    }, [friends])
+
     return (
         <FriendsContext.Provider value={{
-            friends, getFriends
+            friends, filteredFriends, getFriends
         }}>
             {props.children}
         </FriendsContext.Provider>
